Time 10 KDF samples, compute work factor based on preferred runtime, default time factor to 1 second

+++ src/crypto.c
@@ -99,7 +99,7 @@ int sqlcipher_codec_pragma(sqlite3* db, int iDb, Parse *pParse, const char *zLef
   } else
   if(sqlite3StrICmp(zLeft, "cipher_kdf_compute")== 0 && zRight ){
     if(ctx) {
-      char *kdf_iter = sqlite3_mprintf("%ld", sqlcipher_codec_compute_kdf_iter(ctx, atol(zRight)));
+      char *kdf_iter = sqlite3_mprintf("%ld", sqlcipher_codec_compute_kdf_iter(ctx, atof(zRight)));
       codec_vdbe_return_static_string(pParse, "cipher_kdf_compute", kdf_iter);
       sqlcipher_codec_ctx_set_kdf_iter(ctx, atol(kdf_iter), 2);
       sqlite3_free(kdf_iter);
